chore: remove store from components and restructure components folders
