+++ src/index.tsx
@@ -5,9 +5,9 @@ import { Provider } from 'react-redux';
 // Style
 import './index.css';
 // Components
-import App from './components/App/App'
+import App from './App'
 // Store
-import store from './components/store/';
+import store from './store';
 
 
 ReactDOM.render(
