+++ src/store/modules/recorder/index.test.ts
@@ -1,4 +1,4 @@
-import { startRecorder, stopRecorder, recorderReducer } from './';
+import { startRecorder, stopRecorder, recorderReducer } from '.';
 
 describe('Recorder module', () => {
   const initialState = {
